requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-Login==0.6.3
SQLAlchemy==1.4.53
python-dotenv==1.0.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==3.0.1
typing-extensions==4.9.0
